Ações: aplicação de permissões p/ editar/deletar

+++ rap/Acoes/views.py
@@ -2,19 +2,26 @@ from django.shortcuts import render, redirect
 from django.http import HttpResponseRedirect
 from django.views import generic
 from django.views.generic.edit import FormMixin
+
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import LoginRequiredMixin
+
+from django.contrib.auth.decorators import user_passes_test
+from django.contrib.auth.mixins import UserPassesTestMixin
+
 from django.urls import reverse, reverse_lazy
 
 from django.db import transaction
 
 from django.http import HttpResponse 
+from django.core.exceptions import PermissionDenied
 import json
 
 from Acoes.models import Acoes, Midia, MensagemAcoes
 from Acoes.forms import FormNovaMensagem, FormNovaMidia
 from Usuario.models import Usuario
 from Acoes.filters import AcoesFiltro
+
 
 class CriarAcao(LoginRequiredMixin, generic.CreateView):
     model = Acoes
@@ -29,10 +36,14 @@ class CriarAcao(LoginRequiredMixin, generic.CreateView):
     def get_success_url(self):
         return reverse_lazy('acoes:listar_usuario', kwargs={'pk': self.request.user.id})
 
-class EditarAcao(LoginRequiredMixin, generic.UpdateView):
+class EditarAcao(LoginRequiredMixin, UserPassesTestMixin, generic.UpdateView):
     model = Acoes
     fields = ['titulo', 'tipo', 'data_inicio', 'data_fim', 'local', 'descricao', 'status']
     template_name = "Acoes/editar.html"
+
+    def test_func(self):
+        acao = Acoes.objects.get(pk=int(self.kwargs['pk']))
+        return self.request.user.id == acao.responsavel.id
 
     def form_valid(self, form):
         usuario = Usuario.objects.get(id=self.request.user.id)
@@ -53,32 +64,33 @@ class EditarAcao(LoginRequiredMixin, generic.UpdateView):
 def editar_midia(request, pk):
 
     acao = Acoes.objects.get(id=pk)
+    if (request.user.id == acao.responsavel.id):
+        if (request.method == 'POST'):
+            form = FormNovaMidia(request.POST, request.FILES)
+            if form.is_valid():
+                Midia.objects.create(acao=acao, midia=form.cleaned_data.get('midia'))
 
-    if (request.method == 'POST'):
-        form = FormNovaMidia(request.POST, request.FILES)
-        if form.is_valid():
-            Midia.objects.create(acao=acao, midia=form.cleaned_data.get('midia'))
-    #     form = FormNovaMidia(data=request.POST, files=request.FILES)
-    #     if form.is_valid():
-    #         for file in request.FILES.getlist('listing_images'):
-    #             Midia.objects.create(acao=acao, midia=file)
-    # else:
-    midias_acao = Midia.objects.filter(acao__id = pk)
+        midias_acao = Midia.objects.filter(acao__id = pk)
 
-    informacoes =  {
-        'acao': acao,
-        'midias_acao': midias_acao,
-        'form_nova_midia': FormNovaMidia()
-    }
+        informacoes =  {
+            'acao': acao,
+            'midias_acao': midias_acao,
+            'form_nova_midia': FormNovaMidia()
+        }
 
-    return render(request, "Acoes/editar_midias.html", informacoes)
+        return render(request, "Acoes/editar_midias.html", informacoes)
+    else:
+        raise PermissionDenied()
 
 @login_required
 def deletar_midia(request,pk):
     midia=Midia.objects.get(id=pk)
-    acao=midia.acao
-    midia.delete()
-    return redirect('acoes:editar_midia', pk=acao.pk)
+    if (request.user.id == midia.acao.responsavel.id):
+        acao=midia.acao
+        midia.delete()
+        return redirect('acoes:editar_midia', pk=acao.pk)
+    else:
+        raise PermissionDenied()
 
 class EditarMidia(LoginRequiredMixin, generic.DetailView):
     model = Midia
@@ -146,18 +158,26 @@ class DetalhesAcao(LoginRequiredMixin, FormMixin, generic.DetailView):
 @login_required
 def deletar(request, pk):
     acao = Acoes.objects.get(pk=pk)
-    acao.deletada = True
-    acao.save()
-
-    return redirect('acoes:listar')
+    if (request.user.id == acao.responsavel.id):
+        acao.deletada = True
+        acao.save()
+        return redirect('acoes:listar')
+    else:
+        raise PermissionDenied()
 
 @login_required
 def deletar_mensagem(request, pk):
     mensagem = MensagemAcoes.objects.get(id=pk)
-    acao_id = mensagem.acao.id
-    mensagem.delete()
+    if (request.user.id == mensagem.usuario.id or 
+        request.user.id == mensagem.acao.responsavel.id or
+        request.user.is_superuser or
+        (mensagem.mensagem_original is not None and request.user.id == mensagem.mensagem_original.usuario.id)):
+        acao_id = mensagem.acao.id
+        mensagem.delete()
 
-    return HttpResponseRedirect(reverse('acoes:detalhes', kwargs={'pk': acao_id}))
+        return HttpResponseRedirect(reverse('acoes:detalhes', kwargs={'pk': acao_id}))
+    else:
+        raise PermissionDenied()
     
 class ListarTodasAcoes(LoginRequiredMixin, generic.ListView):
     model = Acoes
@@ -179,11 +199,14 @@ class ListarTodasAcoes(LoginRequiredMixin, generic.ListView):
         context['exibir_todos'] = True
         return context
     
-class ListarAcoesUsuario(LoginRequiredMixin, generic.ListView):
+class ListarAcoesUsuario(LoginRequiredMixin, UserPassesTestMixin, generic.ListView):
     model = Acoes
     template_name = 'Acoes/listar.html'
     context_object_name = 'lista_acoes'
     paginate_by = 10
+
+    def test_func(self):
+        return self.request.user.id == int(self.kwargs['pk'])
 
     def get_queryset(self):
         qs = super().get_queryset()
@@ -202,9 +225,17 @@ class ListarAcoesUsuario(LoginRequiredMixin, generic.ListView):
 @login_required
 def alterar_status_acao(request, pk):
     acao = Acoes.objects.get(pk=pk)
-    acao.status = not acao.status
-    acao.save()
-    return finalizar_requisicao_api(acao.status)
+    print(request.user.id)
+    print(acao.responsavel.id)
+    if (request.user.id == acao.responsavel.id):
+        acao.status = not acao.status
+        acao.save()
+        return HttpResponse(
+            json.dumps(acao.status),
+            content_type="application/json"
+        )
+    else:
+        raise PermissionDenied()
 
 @login_required
 def finalizar_requisicao_api(response_data):
